CORNINSPEC-INLINE-BE/CORNINSPEC-INLINE-PREDICT-SERVICE/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.0
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
ecdsa==0.18.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
flask==2.0.3
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==7.31.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
limits==3.10.1
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.26.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.42
urllib3==1.26.6
uvicorn==0.29.0
wcwidth==0.2.13
wrapt==1.16.0
